fix(email): avoid abs() overflow in health-poll hash bucketing
Cast hashtext to bigint before abs() so a hash of i32::MIN can't trip
Postgres' "integer out of range" error and abort the bucketing query.
Addresses PR #4132 review.

+++ rust/cloud-storage/email_refresh_handler/src/handler.rs
@@ -58,7 +58,7 @@ async fn send_health_check_messages(ctx: &context::Context) -> Result<(), Error>
         WHERE
             is_sync_active = TRUE
             AND provider = $1
-            AND (abs(hashtext(id::text)) % $2::int4) = $3::int4
+            AND (abs(hashtext(id::text)::bigint) % $2::int4) = $3::int4
         "#,
         provider_filter as _,
         interval_hours,
